结算功能-int 更新 double --未更新impl

+++ hrms-jscx/src/main/java/jssvc/hrms/dao/SettlementMapper.java
@@ -20,7 +20,7 @@ public interface SettlementMapper {
 
     List<Settlement> selectSettlement(SettlementSearchFilter filter);
 
-    int selectPost(String dah);
+    double selectPost(String dah);
 
     int selectSettlementCount(SettlementSearchFilter filter);
